refactor(gui): move ThumbnailCache to extra file
